requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.1
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.7.0
